Fixed file open bug, removed BRS references
Fixed a bug that kept opening background opacity files by not closing
the file descriptor. It some cases this would lead to a crash when
running with many tasks per process. Also removed all references and
code to the netcdf BRS file, which is no longer used and is confusing.